Improved Frames API, added CIRF.
Using __getattr__ is slow for FrameProxy, so now inherits from AbstractFrame and implements proxy properties directly.

The FrameProxy now uses a function to register the factory function, which is cleaner. I didn't want to have

@FrameProxy
def GCRF():
    return Frame(...)

since that isn't as clear reading the source that GCRF is a FrameProxy.